you need to configure the style
I have tested the repository and found that you need to edit the project options to include the style reference. Otherwise the application will show an message box.

This is because I chose to not publish the .dproj file.
The generated .dproj will not have the style reference.

Add the style from within the Delphi IDE and all is good.

The githup repo seems to work!